(issue14) Move all services to respective app/services folder
