+++ app/services/chat_service.py
