+++ tests/app/services/planner_service_test.py
@@ -1,20 +1,12 @@
-import sys
-import os
 import pytest
 from unittest.mock import patch
-
-sys.path.insert(
-    0,
-    os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "..")),
-)
-
-from agents.services.planner_service import PlannerService
+from app.services.planner_service import PlannerService
 
 
 @pytest.fixture(scope="function")
 def planner_service_with_mocked_chat():
     with patch(
-        "agents.services.planner_service.UserProxyAgent.initiate_chat"
+        "app.services.planner_service.UserProxyAgent.initiate_chat"
     ) as MockedInitiateChat:
         MockedInitiateChat.return_value = (
             None  # No return value as we're mocking the method
@@ -31,7 +23,8 @@ def planner_service_with_mocked_chat():
 
 
 def test_get_service(planner_service_with_mocked_chat):
-    assert planner_service_with_mocked_chat.get_service("test_service") == "TestService"
+    assert planner_service_with_mocked_chat.get_service(
+        "test_service") == "TestService"
 
 
 def test_get_parameters(planner_service_with_mocked_chat):
@@ -42,8 +35,10 @@ def test_get_parameters(planner_service_with_mocked_chat):
 
 def test_ask_planner(planner_service_with_mocked_chat):
     with patch(
-        "agents.services.planner_service.UserProxyAgent.last_message"
+        "app.services.planner_service.UserProxyAgent.last_message"
     ) as MockedLastMessage:
-        MockedLastMessage.return_value = {"content": "Test response from planner"}
-        response = planner_service_with_mocked_chat.ask_planner("buy groceries")
+        MockedLastMessage.return_value = {
+            "content": "Test response from planner"}
+        response = planner_service_with_mocked_chat.ask_planner(
+            "buy groceries")
         assert response == "Test response from planner"
